requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
Pillow==10.2.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.31.0